% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-utils.R
\name{angstrom_to_keV}
\alias{angstrom_to_keV}
\title{Convert angstrom to keV.}
\usage{
angstrom_to_keV(x, x_unit = "angstrom")
}
\arguments{
\item{x}{value}

\item{x_unit}{unit of value (default="angstrom")}
}
\value{
photon energy in keV units
}
\description{
Convert angstrom to keV.
}
