% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossmatch.R
\name{cross_match_r}
\alias{cross_match_r}
\title{Catalog cross-matching.}
\usage{
cross_match_r(x, y, column_src, column_coord, r_arcsec = 30)
}
\arguments{
\item{x}{data.frame with coulumn_src and column_coord}

\item{y}{data.frame with coulumn_src and column_coord}

\item{column_src}{name of the source}

\item{column_coord}{skycoord column}

\item{r_arcsec}{maximum angular distance in arcsec (default is 30 arcsec)}
}
\value{
data.frame
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Deprecated, use instead \link{angsep_left_join}.
}
