% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{distance_haversine}
\alias{distance_haversine}
\title{Angular distance}
\usage{
distance_haversine(ra1, dec1, ra2, dec2)
}
\arguments{
\item{ra1, dec1}{equatorial coordinates of object 1 in degrees}

\item{ra2, dec2}{equatorial coordinates of object 2 in degrees}
}
\value{
distance in arcsec
}
\description{
Angular distance
}
\seealso{
\code{\link[=distance_vincenty]{distance_vincenty()}}
}
