% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xray_utils.R
\name{energy_conv_factor}
\alias{energy_conv_factor}
\title{Energy Conversion Factor}
\usage{
energy_conv_factor(E1, E2, pho_ind = 2)
}
\arguments{
\item{E1}{energy min in keV}

\item{E2}{energy max in keV}

\item{pho_ind}{Photon index, 2.0 default.}
}
\value{
numeric
}
\description{
Convert energy flux to photon flux
}
\details{
see https://hea-www.harvard.edu/~pgreen/figs/xray_cheat.pdf
}
\examples{

energy_conv_factor(0.5, 8.0, pho_ind = 1.7)

}
