% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-utils.R
\name{fix_header_ps1}
\alias{fix_header_ps1}
\title{Fix ps1 header}
\usage{
fix_header_ps1(header)
}
\arguments{
\item{header}{FITSio --> obj$hdr}
}
\value{
fixed_header
}
\description{
Fix ps1 header
}
\examples{
\dontrun{
f <- FITSio::readFITS("file.fits")
fix_header(f$hdr)
}
}
