% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{format.astro_radec}
\alias{format.astro_radec}
\title{Title}
\usage{
\method{format}{astro_radec}(x, formatter = formatter_hmsdms, nsec = 1, color = F, isPlain = F, ...)
}
\arguments{
\item{color}{print color?}

\item{isPlain}{units - white spaces}
}
\value{

}
\description{
Title
}
