% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{formatter_hmsdms}
\alias{formatter_hmsdms}
\title{Title}
\usage{
formatter_hmsdms(x, ra_or_dec, nsec = 1, color = FALSE, ...)
}
\arguments{
\item{color}{print with color?}

\item{...}{}
}
\value{

}
\description{
Title
}
