% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xray_pointrange.R
\name{geom_xray_pointrange}
\alias{geom_xray_pointrange}
\title{Title}
\usage{
geom_xray_pointrange(
  mapping = NULL,
  data = NULL,
  stat = "XrayPointrange",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  upper_nsigma = 2,
  min_y = NULL,
  ...
)
}
\arguments{
\item{min_y}{}

\item{...}{invisible args}
}
\value{

}
\description{
Title
}
