% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xray_ratio.R
\name{geom_xray_ratio}
\alias{geom_xray_ratio}
\title{Title}
\usage{
geom_xray_ratio(
  mapping = NULL,
  data = NULL,
  stat = "XrayRatio",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{invisible args}
}
\value{

}
\description{
Title
}
