% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xray_stepcount.R
\name{geom_xray_stepcount}
\alias{geom_xray_stepcount}
\title{Title}
\usage{
geom_xray_stepcount(
  mapping = NULL,
  data = NULL,
  stat = "XrayStepCount",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{invisible args}
}
\value{

}
\description{
Title
}
