% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_date.R
\name{jd}
\alias{jd}
\title{Convert greg to JD astronomical}
\usage{
jd(year, month, day, hour, min, sec)
}
\arguments{
\item{year}{year}

\item{sec}{}
}
\description{
Convert greg to JD astronomical
}
\examples{
\dontrun{
jd1 <- jd(2000, 1, 1, 12, 0, 0)
jd1
}
}
