% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julian_date.R
\name{jd_fromdate}
\alias{jd_fromdate}
\title{Convert greg to JD astronomical. At first it converts to UTC.}
\usage{
jd_fromdate(date)
}
\arguments{
\item{date}{}
}
\value{

}
\description{
Convert greg to JD astronomical. At first it converts to UTC.
}
\examples{
\dontrun{
d0 <- ymd_hms('2000/01/01 15:00:00', tz='Europe/Moscow')
jd1 <- jd_fromdate(d0)
jd1
}
}
