% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex_table.R
\name{latex_range}
\alias{latex_range}
\title{Make latex notation for best value and confidence boundaries.}
\usage{
latex_range(best, min, max, ndigits = 0, na = "-")
}
\arguments{
\item{best}{best value}

\item{min}{lower boundary value}

\item{max}{upper boundary value}

\item{ndigits}{number of digits after decimal point}

\item{na}{show NA value as na. default na='-'}
}
\value{
character vector
}
\description{
Make latex notation for best value and confidence boundaries.
}
\examples{
x <- c(NA,1:10, NA)
x <- x + rnorm(length(x))
err <- rnorm(length(x))

latex_range(best = x, min = x - err, max = x + err, ndigits = 2)

latex_range(best = x, min = x - err, max = x + 2*err, ndigits = 2, na='')

}
