% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.R
\name{loglikelihood_chi}
\alias{loglikelihood_chi}
\title{Likelihood for Gaussian data}
\usage{
loglikelihood_chi(obs, model)
}
\arguments{
\item{obs}{observed data rates}

\item{model}{the values of the predicted data rates based on the model

see https://heasarc.gsfc.nasa.gov/xanadu/xspec/manual/XSappendixStatistics.html}
}
\description{
Likelihood for Gaussian data
}
