% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh.R
\name{nh}
\alias{nh}
\title{Calculate the HI Column Density for a Sky Position}
\usage{
nh(ra = NULL, dec = NULL, r_arcmin = 10, showInfo = FALSE)
}
\arguments{
\item{ra}{RA in degrees}

\item{dec}{DEC in degrees}

\item{r_arcmin}{search radius in arcmin}

\item{showInfo}{show additional information ? Default=false}
}
\value{
weighted nH
}
\description{
nH is calculated using all points (from https://vizier.u-strasbg.fr site.) within `r_arcmin`
}
\examples{
\dontrun{
 xrayr::nh(213.2, -65.4, r_arcmin = 30, showInfo = T)
}
}
