% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optical-utils.R
\name{plot_guidance}
\alias{plot_guidance}
\title{Make guidance image}
\usage{
plot_guidance(
  image,
  hdr,
  pixel_scale = NULL,
  r_arcsec = NULL,
  radec = NULL,
  colors = NULL,
  label = NULL,
  label_size = 2,
  circle_lwd = 1,
  scale_lwd = 1,
  scale_txt_size = 1,
  compass_lwd = 1,
  level.lo = 0.6,
  level.hi = 0.99,
  arrow_radec = NULL,
  arrow_size = 1,
  arrow_lwd = 1,
  arrow_color = "red",
  show_wcs_compass = TRUE,
  show_image_scale = TRUE
)
}
\arguments{
\item{image}{image of class matrix...}

\item{hdr}{header with keywords...}

\item{pixel_scale}{scale in units arcsec/pixel; for ps1dr2 0.25"/pixel}

\item{r_arcsec}{vector of radiuses of each circle}

\item{radec}{astro_radec vector}

\item{colors}{colors of circles}

\item{label}{main image label}

\item{label_size}{size of main label}

\item{circle_lwd}{linewidth of circle regions}

\item{scale_lwd}{linewidth of scale line}

\item{scale_txt_size}{text size of scale}

\item{compass_lwd}{linewidth of compass lines}

\item{level.lo}{level of image grayscale; default 0.6}

\item{level.hi}{level of image grayscale; default 0.99}

\item{arrow_color}{}

\item{show_wcs_compass}{If FALSE, do not show WCS compass on image. If TRUE show the WCS compass...}

\item{show_image_scale}{If FALSE, do not show scale legend on image. If TRUE show the legend of scale on image...}
}
\description{
Make guidance image
}
\examples{
\dontrun{
par(oma = c(0, 0, 0, 0), mar = c(0, 0, 0, 0))
plot_guidance2(f$imDat,
hdr = fix_header_ps1(f$hdr),
radec = c(radec("J010744.2+574428"),
          radec("J010745.2+574432")),
r_arcsec = c(30, 5),
colors = c('blue', 'red'),
pixel_scale = 0.25,label = "SRGA",
show_image_scale = TRUE,
show_wcs_compass = TRUE)
}
}
