% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{radec}
\alias{radec}
\title{Title}
\usage{
radec(radec = character())
}
\arguments{
\item{radec}{in "hms dms"}
}
\description{
Title
}
\examples{
radec('12 34 56 -76  54 3.210')
radec('12 34 56 -76  54 3.210') \%>\% tibble::tibble()
}
