% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qdp.R
\name{read_qdp}
\alias{read_qdp}
\title{Title}
\usage{
read_qdp(
  filename,
  telescope_names = NULL,
  col_names = c("energy", "err_energy", "flux_obs", "err_obs", "flux_model")
)
}
\arguments{
\item{filename}{name of the file}

\item{telescope_names}{vector of the telescope names}

\item{col_names}{default: c('energy', 'err_energy', 'flux_obs', 'err_obs', 'flux_model')}
}
\value{

}
\description{
Title
}
