% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{separation.astro_radec}
\alias{separation.astro_radec}
\title{Angular distance between two points}
\usage{
\method{separation}{astro_radec}(x, y, metric = distance_vincenty)
}
\arguments{
\item{metric}{\code{\link[=distance_vincenty]{distance_vincenty()}} (default) or \code{\link[=distance_haversine]{distance_haversine()}}}
}
\description{
Angular distance between two points
}
