% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xray_pointrange.R
\name{stat_xray_pointrange}
\alias{stat_xray_pointrange}
\title{Title}
\usage{
stat_xray_pointrange(
  mapping = NULL,
  data = NULL,
  geom = "xray_pointrange",
  position = "identity",
  show.legend = NA,
  outliers = TRUE,
  inherit.aes = TRUE,
  upper_nsigma = 2,
  min_y = NULL,
  ...
)
}
\arguments{
\item{min_y}{}

\item{...}{invisible args}
}
\description{
Title
}
