% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xray_ratio.R
\name{stat_xray_ratio}
\alias{stat_xray_ratio}
\title{Title}
\usage{
stat_xray_ratio(
  mapping = NULL,
  data = NULL,
  geom = "xray_ratio",
  position = "identity",
  show.legend = NA,
  outliers = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{mapping}

\item{data}{data}

\item{geom}{geom}

\item{position}{position}

\item{show.legend}{show.legend}

\item{outliers}{outliers}

\item{inherit.aes}{ingherit.aes}

\item{...}{invisible args}
}
\description{
Title
}
