% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_xray_stepcount.R
\name{stat_xray_stepcount}
\alias{stat_xray_stepcount}
\title{Title}
\usage{
stat_xray_stepcount(
  mapping = NULL,
  data = NULL,
  geom = "xray_stepcount",
  position = "identity",
  show.legend = NA,
  outliers = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{...}{invisible args}
}
\description{
Title
}
