% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{staticSuperResolution}
\alias{staticSuperResolution}
\title{Static super resolution algorithms.}
\usage{
staticSuperResolution(inputImage, expansionFactor, method = "sharpen")
}
\arguments{
\item{inputImage}{image to be expanded}

\item{expansionFactor}{float value}

\item{method}{one of sharpen, pysr, neurenhance}
}
\value{
super resolution image
}
\description{
Implements several different strategies for super-resolution from static data.
}
\examples{

library( magrittr )
library( ANTsRCore )
library( ANTsR )
temp1 = resampleImage( ri(1), 8 )
sup = staticSuperResolution( temp1, 2 )

}
\author{
Avants BB
}
