% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_xray_plot}
\alias{theme_xray_plot}
\title{Title}
\usage{
theme_xray_plot(is_transparent = T, ...)
}
\arguments{
\item{is_transparent}{transparent background}

\item{...}{invisible args}
}
\description{
Title
}
