% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{timeSeriesMotionCorrect}
\alias{timeSeriesMotionCorrect}
\title{Suggested parameters for deformable motion correction for a time series.}
\usage{
timeSeriesMotionCorrect(fixedStaticImage, movingTimeSeries, totalSigma,
  flowSigma, regIterations = c(50, 50, 10), ...)
}
\arguments{
\item{fixedStaticImage}{fixed static image}

\item{movingTimeSeries}{time series image to map to the fixed space}

\item{totalSigma}{regularization for total deformation field}

\item{flowSigma}{regularization for update deformation field}

\item{regIterations}{deformable registration iterations / multi-resolution parameters}

\item{...}{parameters to pass to antsRegistration}
}
\value{
deformed images
}
\description{
This function, by default, regularizes the total deformation field to produce a
a constrained diffeomorphic motion correction of a time series.  Full custom
by passing the same parameters that one would pass to \code{antsRegistration}.
}
\examples{

library( magrittr )
library( ANTsRCore )
library( ANTsR )
temp1 = resampleImage( ri(1), 4 )
temp2 = resampleImage( ri(2), 4 )
temp3 = resampleImage( ri(1), 2 )
tarspace = makeImage( c( dim( temp1 ), 2 ), spacing = c( 4, 4, 1 ) )
pdcr = mergeListToNDImage( tarspace, list( temp1, temp2 ) )
reg = timeSeriesMotionCorrect( temp3, pdcr )

}
\author{
Avants BB
}
