% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{timeSeriesSuperResolution}
\alias{timeSeriesSuperResolution}
\title{Time series super resolution algorithms.}
\usage{
timeSeriesSuperResolution(fixedStaticImage, mappedTimeSeries,
  method = "average", weight = 0.5)
}
\arguments{
\item{fixedStaticImage}{fixed target image}

\item{mappedTimeSeries}{pre-aligned to the fixed space}

\item{method}{one of average, sharpen, pysr, neurenhance, majon}

\item{weight}{generic weight term}
}
\value{
super resolution image
}
\description{
Implements several different strategies for super-resolution from time series data.
}
\examples{

library( magrittr )
library( ANTsRCore )
library( ANTsR )
temp1 = resampleImage( ri(1), 8 )
temp2 = resampleImage( ri(1), 8 )
temp3 = resampleImage( ri(1), 4 )
tarspace = makeImage( c( dim( temp1 ), 2 ) , spacing = c( 8, 8, 1 ) )
pdcr = mergeListToNDImage( tarspace, list( temp1, temp2 ) )
reg = timeSeriesMotionCorrect( temp3, pdcr )
sup = timeSeriesSuperResolution( temp3, reg )

}
\author{
Avants BB
}
