% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{to_galactic.astro_radec}
\alias{to_galactic.astro_radec}
\title{Convert radec to galactic coords.}
\usage{
\method{to_galactic}{astro_radec}(x, obs_epoch = "J2000")
}
\arguments{
\item{x}{\code{astro_radec} object}

\item{epoch}{"J2000" (default) or "B1950"}
}
\value{
list(l=longitude\link{deg}, b = latitude\link{deg})
}
\description{
Convert radec to galactic coords.
}
