% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xray_spec}
\alias{xray_spec}
\title{Xray spec which used in vegnette}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 25 rows and 6 columns.
}
\usage{
data(xray_spec)
}
\description{
Xray spec which used in vegnette
}
\keyword{datasets}
