% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{xrayMotionCorrect}
\alias{xrayMotionCorrect}
\title{Suggested parameters for deformable intrasubject xray registration.}
\usage{
xrayMotionCorrect(fixed, moving, totalSigma, flowSigma,
  regIterations = c(50, 50, 10), ...)
}
\arguments{
\item{fixed}{fixed target image}

\item{moving}{image to map to the fixed space}

\item{totalSigma}{regularization for total deformation field}

\item{flowSigma}{regularization for update deformation field}

\item{regIterations}{deformable registration iterations / multi-resolution parameters}

\item{...}{parameters to pass to antsRegistration}
}
\value{
deformed image
}
\description{
This function, by default, regularizes the total deformation field to produce a
a constrained diffeomorphic mapping.  Full custom parameters are possible
by passing the same parameters that one would pass to \code{antsRegistration}.
}
\examples{

library( magrittr )
i1 = ANTsRCore::ri( 1 ) \%>\% ANTsRCore::resampleImage( 4 )
i2 = ANTsRCore::ri( 2 ) \%>\% ANTsRCore::resampleImage( 4 )
reg = xrayMotionCorrect( i1, i2 )

}
\author{
Avants BB
}
