% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrayr-package.R
\docType{package}
\name{xrayr-package}
\alias{xrayr}
\alias{xrayr-package}
\title{xrayr: Tools for xray astronomy}
\description{
Tools for xray astronomy.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/uskovgs/xrayr/}
  \item Report bugs at \url{{{ BugReports }}}
}

}
\author{
\strong{Maintainer}: Grigory Uskov \email{uskov.russia@gmail.com}

}
\keyword{internal}
