% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xspec_output}
\alias{xspec_output}
\title{Example of xray output}
\format{
An object of class \code{character} of length 89.
}
\usage{
data(xray_output)
}
\description{
Example of xray output
}
\keyword{datasets}
