% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspec.R
\name{xspec_parse_conf_errors}
\alias{xspec_parse_conf_errors}
\title{Parse xspec output: confidence errors}
\usage{
xspec_parse_conf_errors(xspec_lines)
}
\arguments{
\item{xspec_lines}{xspec output}
}
\value{
tibble
}
\description{
Parse xspec output: confidence errors
}
