% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspec.R
\name{xspec_parse_model}
\alias{xspec_parse_model}
\title{Parse xspec output: model}
\usage{
xspec_parse_model(xspec_lines)
}
\arguments{
\item{xspec_lines}{xspec output}
}
\value{
character string
}
\description{
Parse xspec output: model
}
