% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspec.R
\name{xspec_parse_output}
\alias{xspec_parse_output}
\title{Parse xspec output: model, fit params and conf errors}
\usage{
xspec_parse_output(xspec_lines)
}
\arguments{
\item{xspec_lines}{xspec output}
}
\value{
list with components: model, fit_params, conf_errors
}
\description{
See vignette for more details
}
