% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xspec.R
\name{xspec_parse_params}
\alias{xspec_parse_params}
\title{Parse xspec output: fit params}
\usage{
xspec_parse_params(xspec_lines)
}
\arguments{
\item{xspec_lines}{xspec output}
}
\value{
tibble
}
\description{
Parse xspec output: fit params
}
