% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nh.R
\name{nh}
\alias{nh}
\title{Calculate the HI Column Density for a Sky Position}
\usage{
nh(ra, dec, r_arcmin = 10, showInfo = T)
}
\arguments{
\item{ra}{RA in degrees}

\item{dec}{DEC in degrees}

\item{r_arcmin}{search radius in arcmin}

\item{showInfo}{show additional information ? Default=true}
}
\value{

}
\description{
nH calculated using all points (from https://vizier.u-strasbg.fr site.) within \code{r_arcmin}
}
\examples{
\dontrun{
nh(213.162500, -65.392778, r_arcmin = 30, showInfo = T)
}
}
