% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/src.R
\name{timeSeriesSuperResolution}
\alias{timeSeriesSuperResolution}
\title{Time series super resolution algorithms.}
\usage{
timeSeriesSuperResolution(fixedStaticImage, mappedTimeSeries,
  method = "average", param1, param2)
}
\arguments{
\item{fixedStaticImage}{target image space}

\item{mappedTimeSeries}{pre-aligned to the fixed space, not necessarily the same resolution}

\item{method}{one of average, sharpen, majon}

\item{param1}{generic weight term for sharpen or mesh size for majon}

\item{param2}{number of levels for majon}
}
\value{
super resolution image
}
\description{
Implements several different strategies for super-resolution from time series data.
}
\examples{

library( magrittr )
library( ANTsRCore )
library( ANTsR )
temp1 = resampleImage( ri(1), 8 )
temp2 = resampleImage( ri(1), 8 )
temp3 = resampleImage( ri(1), 4 )
tarspace = makeImage( c( dim( temp1 ), 2 ) , spacing = c( 8, 8, 1 ) )
pdcr = mergeListToNDImage( tarspace, list( temp1, temp2 ) )
sup = timeSeriesSuperResolution( temp3, pdcr )

}
\author{
Avants BB
}
