% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radec.R
\name{to_galactic.astro_radec}
\alias{to_galactic.astro_radec}
\title{Convert radec to galactic coords.}
\usage{
\method{to_galactic}{astro_radec}(x, obs_epoch = "J2000")
}
\arguments{
\item{x}{radec obj}

\item{epoch}{"J2000" (default) or "B1950"}
}
\value{

}
\description{
See formula from www.atnf.csiro.au/people/Tobias.Westmeier/tools_coords.php
}
